requirements.txt
streamlit
python-dotenv
langchain
langchain-openai
langchain-community
openai
pypdf
pydantic
gitpython
faiss-cpu
pandasstre
